Renamed InMemoryItemStore file
